# apps/api/internal/spec/openapi.yaml
openapi: 3.0.3
info:
  title: UFO Hub API
  version: 0.7.0
  license:
    name: BSD-3-Clause
    url: https://opensource.org/license/bsd-3-clause
  description: >
    Public beta Hub HTTP API for UFO (Unified Fleet Orchestrator). Endpoints and
    payloads may still change before 1.0; match clients to tagged releases and
    their release notes.
    Three surfaces: auth (public), the tenant/UI surface (httpOnly access JWT cookie
    or bearer token), and the rover surface (per-rover connection token). Source of truth for the
    protocol; maintained by hand alongside the code (generated clients are a follow-up).
servers:
  - url: /
security: []
...
paths:
...
  /v1/usage:
    get:
      tags: [ui]
      operationId: getUsage
      summary: Fleet and mission usage totals for a calendar period
      security: [{ accessCookie: [] }, { userBearer: [] }]
      parameters:
        - name: fleet_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: period
          in: query
          schema: { type: string, enum: [calendar_week, calendar_month], default: calendar_week }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/operations/{id}:
    get:
      tags: [ui]
      operationId: getOperation
      summary: Operation detail (comments, runs, relations, source actions)
      security: [{ accessCookie: [] }, { userBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [ui]
      operationId: patchOperation
      summary: Update an operation (title, body, status, assignee, tags, ...)
      security: [{ accessCookie: [] }, { userBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperationPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/missions/{id}:
    get:
      tags: [ui]
      operationId: getMission
      summary: Get a mission by public id
      security: [{ accessCookie: [] }, { userBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Mission" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [ui]
      operationId: updateMission
      summary: Rename a mission and/or its key
      security: [{ accessCookie: [] }, { userBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMissionRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: ufo_access
      description: "Browser auth: httpOnly access JWT cookie."
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access JWT for non-browser clients.
    enrollmentCode:
      type: http
      scheme: bearer
      description: Enrollment code from the Rovers panel; exchanged once for a connection token.
    roverConnectionToken:
      type: http
      scheme: bearer
      description: Per-rover connection token returned by enrollment.
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Public resource id.
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Bad request
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Not a member of the fleet
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Already exists
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      properties: { error: { type: string } }
...
    Operation:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        status: { type: string, enum: [backlog, todo, in_progress, in_review, done, blocked, canceled] }
        active_run_status: { type: string, enum: ["", queued, accepted, starting, running], description: "Current in-flight run status, empty when none" }
        mission_id: { type: string, format: uuid, description: The mission this operation belongs to (required) }
        sequence: { type: integer, description: Per-mission running number; displayed as <mission.key>-<sequence> }
        priority: { type: integer, description: 0 none → 4 urgent }
        assignee_type: { type: string, enum: [pilot, user, crew], nullable: true }
        assignee_id: { type: string, format: uuid, nullable: true, description: Referenced user/crew public id (null when assignee is a pilot) }
        assignee_pilot_kind: { type: string, nullable: true, description: Pilot kind when assignee_type=pilot (e.g. codex) }
        required_tags: { type: array, items: { type: string }, description: dispatch allow list (rover must have all) }
        excluded_tags: { type: array, items: { type: string }, description: dispatch deny list (rover must have none; deny wins) }
        labels: { type: array, items: { $ref: "#/components/schemas/Label" } }
        reactions: { type: array, items: { $ref: "#/components/schemas/Reaction" }, description: only populated in the detail view }
        sub_operation_progress: { type: object, properties: { total: { type: integer }, done: { type: integer }, in_progress: { type: integer }, in_review: { type: integer }, blocked: { type: integer }, pilot_kinds: { type: array, items: { type: string } } }, description: sub-operation progress }
        metadata: { type: object, additionalProperties: true, readOnly: true, description: Operation metadata object for optional feature flags and future fields }
        sub_operations_enabled: { type: boolean, description: Captain may split this operation into sub-operations }
        start_date: { type: string, format: date, nullable: true, description: Planning start date }
        due_date: { type: string, format: date, nullable: true, description: Planning due date }
        main_operation_id: { type: string, format: uuid, nullable: true, description: Main operation public id for a sub-operation; nested sub-operations are rejected }
        orchestrating: { type: boolean, description: captain split this operation into sub-operations and will reconvene to reconcile them }
        archived: { type: boolean }
        created_by: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true, description: Lifecycle timestamp set when work starts }
        finished_at: { type: string, format: date-time, nullable: true, description: Lifecycle timestamp set when work reaches done/canceled }
...
    Relation:
      type: object
      description: A directed link to another operation, shown from this operation's side.
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [blocks, blocked_by, relates, duplicate, duplicated_by] }
        operation: { $ref: "#/components/schemas/OperationReference" }
        created_by: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
    SourceAction:
      type: object
      properties:
        id: { type: string, format: uuid }
        operation_id: { type: string, format: uuid }
        run_id: { type: string, format: uuid, nullable: true }
        rover_id: { type: string, format: uuid, nullable: true }
        kind: { type: string, enum: [apply_to_source, create_source_branch, commit_to_branch, refresh_from_source] }
        status: { type: string, enum: [queued, accepted, succeeded, failed, conflicted] }
        branch_name: { type: string }
        commit_sha: { type: string }
        base_sha: { type: string }
        source_head_sha: { type: string }
        message: { type: string }
        metadata: { type: object, additionalProperties: true }
        created_by: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        accepted_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        author_type: { type: string, enum: [user, pilot, system] }
        author_id: { type: string, format: uuid, nullable: true, description: user public id (null when author is a pilot/system) }
        author_pilot_kind: { type: string, nullable: true, description: pilot kind when author_type=pilot }
        body: { type: string }
        reactions: { type: array, items: { $ref: "#/components/schemas/Reaction" } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Mission:
      type: object
      description: A grouping of operations within a fleet, with a fleet-unique operation-prefix key.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key: { type: string, description: Operation prefix (e.g. MSJ -> MSJ-123); unique within the fleet }
        forge_ids:
          type: array
          items: { type: string, format: uuid }
          description: Forges this mission can use (public ids). Empty means none. Which forge to ship with is set on the routine or operation (forge_key).
        metadata: { type: object, additionalProperties: true, readOnly: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    UpdateMissionRequest:
      type: object
      required: [name, key]
      properties:
        name: { type: string }
        key: { type: string, description: Operation prefix (alphanumeric; unique within the fleet) }
        forge_ids:
          type: array
          items: { type: string, format: uuid }
          description: Replace the forges this mission can use (owner/admin). Omitted leaves the set unchanged.
        metadata: { type: object, additionalProperties: true }
...
    UsageTotals:
      type: object
      required: [runs, total_tokens, cost_micros]
      properties:
        runs: { type: integer, format: int64 }
        total_tokens: { type: integer, format: int64 }
        cost_micros: { type: integer, format: int64 }
        max_runs: { type: integer, format: int64, nullable: true }
        max_tokens: { type: integer, format: int64, nullable: true }
        max_usd_micros: { type: integer, format: int64, nullable: true }
    MissionUsage:
      allOf:
        - { $ref: "#/components/schemas/UsageTotals" }
        - type: object
          required: [id, key, name]
          properties:
            id: { type: string, format: uuid }
            key: { type: string }
            name: { type: string }
    UsageSummary:
      type: object
      required: [period, period_key, start_at, end_at, fleet, missions]
      properties:
        period: { type: string, enum: [calendar_week, calendar_month] }
        period_key: { type: string }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        fleet: { $ref: "#/components/schemas/UsageTotals" }
        missions:
          type: array
          items: { $ref: "#/components/schemas/MissionUsage" }
    Run:
      type: object
      properties:
        id: { type: string, format: uuid }
        operation_id: { type: string, format: uuid }
        pilot: { type: string, pattern: "^[a-z][a-z0-9_-]{0,31}$" }
        status:
          type: string
          enum: [queued, accepted, starting, running, blocked, succeeded, failed, canceled]
        needs_input: { type: boolean, description: pilot ended this run awaiting a human answer }
        metadata: { type: object, additionalProperties: true, readOnly: true }
        usage: { $ref: "#/components/schemas/RunUsage", description: Present on run detail when reported }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    OperationDetail:
      type: object
      properties:
        operation: { $ref: "#/components/schemas/Operation" }
        comments: { type: array, items: { $ref: "#/components/schemas/Comment" } }
        comments_more: { type: boolean }
        runs: { type: array, items: { $ref: "#/components/schemas/Run" } }
        sub_operations: { type: array, items: { $ref: "#/components/schemas/Operation" } }
        relations: { type: array, items: { $ref: "#/components/schemas/Relation" } }
        source_action_available: { type: boolean }
        source_rover_id: { type: string, format: uuid, nullable: true }
        source_actions: { type: array, items: { $ref: "#/components/schemas/SourceAction" } }
        forge_actions: { type: array, items: { $ref: "#/components/schemas/ForgeAction" } }
        pull_requests: { type: array, items: { $ref: "#/components/schemas/PullRequest" } }
...
    ForgeAction:
      type: object
      description: Persisted forge action row (also embedded on operation detail).
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        status: { type: string, enum: [queued, accepted, succeeded, failed, conflicted] }
        provider: { type: string }
        base_url: { type: string }
        repo: { type: string }
        head_branch: { type: string }
        base_branch: { type: string }
        commit_sha: { type: string }
        title: { type: string }
        body: { type: string }
        remote_url: { type: string }
        remote_number: { type: integer, nullable: true }
        result_sha: { type: string }
        message: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    OperationPatch:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        mission_id:
          type: string
          format: uuid
          description: >
            Move the operation (and its sub-operations) to another mission in
            the same fleet. Allocates a new per-mission sequence, clears the
            pilot session so the next run uses the new mission/fleet context,
            and posts a system communication. Rejected while a run is active.
        assignee_type: { type: string, enum: [pilot, user, crew], nullable: true }
        assignee_id: { type: string, nullable: true, description: "user/crew public id, or pilot kind when assignee_type=pilot" }
        status: { type: string, enum: [backlog, todo, in_progress, in_review, done, blocked, canceled] }
        required_tags: { type: array, items: { type: string } }
        excluded_tags: { type: array, items: { type: string } }
        priority: { type: integer, minimum: 0, maximum: 4 }
        start_date: { type: string, format: date, nullable: true }
        due_date: { type: string, format: date, nullable: true }
        main_operation_id: { type: string, format: uuid, nullable: true, description: Main operation public id; must not point to a sub-operation }
        sub_operations_enabled: { type: boolean }
        archived: { type: boolean }
...